redirecting user from editing vidio site & password hash problem solved

+++ src/controllers/videoController.js
@@ -30,7 +30,13 @@ export const see = async (req, res) => {
 export const getEdit = async (req, res) => {
     const { id } = req.params;
     const video = await Video.findById(id);
+    const {
+        user: { _id }
+    } = req.session;
 
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect('/');
+    }
     if (!video) {
         return res.status(404).render('404', { pageTitle: 'Video Not Found' });
     }
@@ -45,6 +51,13 @@ export const postEdit = async (req, res) => {
     const { id } = req.params;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists({ _id: id });
+    const {
+        user: { _id }
+    } = req.session;
+
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect('/');
+    }
 
     if (!video) {
         return res.render('404', { pageTitle: 'Video Not Found' });
@@ -90,6 +103,16 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const {
+        user: { _id }
+    } = req.session;
+    const video = await Video.findById(id);
+    if (!video) {
+        return res.render('404', { pageTitle: 'Video Not Found' });
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect('/');
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect('/');
 };
